+++ Cargo.toml
@@ -55,6 +55,9 @@ cloud-storage-grpc = ["sst_importer/cloud-storage-grpc"]
 cloud-storage-dylib = ["sst_importer/cloud-storage-dylib"]
 pprof-fp = ["pprof/frame-pointer"]
 
+# dummy proxy features to make cargo clippy happy.
+enable-pagestorage = []
+
 # for testing configure propegate to other crates
 # https://stackoverflow.com/questions/41700543/can-we-share-test-utilites-between-crates
 testing = []
