+++ engine_store_ffi/src/engine/mod.rs
@@ -1,6 +1,5 @@
 // Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
 
 pub mod ffihub_impl;
-pub mod ps_engine;
 
 pub use ffihub_impl::*;
