+++ engine_tiflash/src/cached_region_info_manager.rs
@@ -51,6 +51,8 @@ impl CachedRegionInfoManager {
         i ^ (i >> 30) ^ (i >> 60)
     }
 
+    // This is not a trivial default.
+    #[allow(clippy::new_without_default)]
     pub fn new() -> Self {
         let mut cached_region_info = Vec::with_capacity(CACHED_REGION_INFO_SLOT_COUNT);
         for _ in 0..CACHED_REGION_INFO_SLOT_COUNT {
