+++ engine_tiflash/src/engine.rs
@@ -10,19 +10,12 @@ use std::{
     sync::{atomic::AtomicIsize, Arc},
 };
 
-use engine_rocks::{RocksDbVector, RocksEngineIterator, RocksSnapshot};
-use engine_traits::{
-    Checkpointable, Checkpointer, Error, IterOptions, Iterable, KvEngine, Peekable, ReadOptions,
-    Result, SyncMutable,
-};
-use rocksdb::{Writable, DB};
+use engine_rocks::RocksSnapshot;
+use engine_traits::{Checkpointable, Checkpointer, Error, KvEngine, Result};
+use rocksdb::DB;
 
-#[cfg(feature = "enable-pagestorage")]
-use crate::ps_engine::*;
 use crate::{
     proxy_utils::{engine_ext::*, EngineStoreHub},
-    r2e,
-    util::get_cf_handle,
     ProxyEngineExt,
 };
 
@@ -134,104 +127,9 @@ impl KvEngine for RocksEngine {
     }
 }
 
-impl Iterable for RocksEngine {
-    type Iterator = RocksEngineIterator;
-
-    #[cfg(feature = "enable-pagestorage")]
-    fn scan<F>(
-        &self,
-        cf: &str,
-        start_key: &[u8],
-        end_key: &[u8],
-        fill_cache: bool,
-        f: F,
-    ) -> Result<()>
-    where
-        F: FnMut(&[u8], &[u8]) -> Result<bool>,
-    {
-        let mut f = f;
-        self.ps_ext
-            .as_ref()
-            .unwrap()
-            .scan_page(start_key.into(), end_key.into(), &mut f);
-        Ok(())
-    }
-
-    fn iterator_opt(&self, cf: &str, opts: IterOptions) -> Result<Self::Iterator> {
-        self.rocks.iterator_opt(cf, opts)
-    }
-}
-
-#[cfg(not(feature = "enable-pagestorage"))]
-impl Peekable for RocksEngine {
-    type DbVector = RocksDbVector;
-
-    fn get_value_opt(&self, opts: &ReadOptions, key: &[u8]) -> Result<Option<RocksDbVector>> {
-        self.rocks.get_value_opt(opts, key)
-    }
-
-    fn get_value_cf_opt(
-        &self,
-        opts: &ReadOptions,
-        cf: &str,
-        key: &[u8],
-    ) -> Result<Option<RocksDbVector>> {
-        self.rocks.get_value_cf_opt(opts, cf, key)
-    }
-}
-
 impl RocksEngine {
     pub fn do_write(&self, cf: &str, key: &[u8]) -> bool {
         crate::do_write(cf, key)
-    }
-}
-
-#[cfg(not(feature = "enable-pagestorage"))]
-impl SyncMutable for RocksEngine {
-    fn put(&self, key: &[u8], value: &[u8]) -> Result<()> {
-        if self.do_write(engine_traits::CF_DEFAULT, key) {
-            return self.rocks.get_sync_db().put(key, value).map_err(r2e);
-        }
-        Ok(())
-    }
-
-    fn put_cf(&self, cf: &str, key: &[u8], value: &[u8]) -> Result<()> {
-        if self.do_write(cf, key) {
-            let db = self.rocks.get_sync_db();
-            let handle = get_cf_handle(&db, cf)?;
-            return self
-                .rocks
-                .get_sync_db()
-                .put_cf(handle, key, value)
-                .map_err(r2e);
-        }
-        Ok(())
-    }
-
-    fn delete(&self, key: &[u8]) -> Result<()> {
-        if self.do_write(engine_traits::CF_DEFAULT, key) {
-            return self.rocks.get_sync_db().delete(key).map_err(r2e);
-        }
-        Ok(())
-    }
-
-    fn delete_cf(&self, cf: &str, key: &[u8]) -> Result<()> {
-        if self.do_write(cf, key) {
-            let db = self.rocks.get_sync_db();
-            let handle = get_cf_handle(&db, cf)?;
-            return self.rocks.get_sync_db().delete_cf(handle, key).map_err(r2e);
-        }
-        Ok(())
-    }
-
-    fn delete_range(&self, begin_key: &[u8], end_key: &[u8]) -> Result<()> {
-        // do nothing
-        Ok(())
-    }
-
-    fn delete_range_cf(&self, cf: &str, begin_key: &[u8], end_key: &[u8]) -> Result<()> {
-        // do nothing
-        Ok(())
     }
 }
