+++ engine_tiflash/src/ps_engine/mod.rs
@@ -4,8 +4,14 @@
 mod engine;
 #[cfg(feature = "enable-pagestorage")]
 mod ps_db_vector;
+mod ps_log_engine;
+#[cfg(feature = "enable-pagestorage")]
+pub(crate) mod ps_write_batch;
 
 #[cfg(feature = "enable-pagestorage")]
 pub use engine::*;
 #[cfg(feature = "enable-pagestorage")]
 pub use ps_db_vector::*;
+pub use ps_log_engine::*;
+#[cfg(feature = "enable-pagestorage")]
+pub use ps_write_batch::*;
