+++ engine_tiflash/src/rocks_engine/engine.rs
@@ -1,0 +1,84 @@
+// Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
+
+#![allow(unused_variables)]
+
+use engine_rocks::{RocksDbVector, RocksEngineIterator};
+use engine_traits::{IterOptions, Iterable, Peekable, ReadOptions, Result, SyncMutable};
+use rocksdb::Writable;
+
+use crate::{r2e, util::get_cf_handle, RocksEngine};
+
+impl Iterable for RocksEngine {
+    type Iterator = RocksEngineIterator;
+
+    fn iterator_opt(&self, cf: &str, opts: IterOptions) -> Result<Self::Iterator> {
+        self.rocks.iterator_opt(cf, opts)
+    }
+}
+
+#[cfg(not(feature = "enable-pagestorage"))]
+impl Peekable for RocksEngine {
+    type DbVector = RocksDbVector;
+
+    fn get_value_opt(&self, opts: &ReadOptions, key: &[u8]) -> Result<Option<RocksDbVector>> {
+        self.rocks.get_value_opt(opts, key)
+    }
+
+    fn get_value_cf_opt(
+        &self,
+        opts: &ReadOptions,
+        cf: &str,
+        key: &[u8],
+    ) -> Result<Option<RocksDbVector>> {
+        self.rocks.get_value_cf_opt(opts, cf, key)
+    }
+}
+
+#[cfg(not(feature = "enable-pagestorage"))]
+impl SyncMutable for RocksEngine {
+    fn put(&self, key: &[u8], value: &[u8]) -> Result<()> {
+        if self.do_write(engine_traits::CF_DEFAULT, key) {
+            return self.rocks.get_sync_db().put(key, value).map_err(r2e);
+        }
+        Ok(())
+    }
+
+    fn put_cf(&self, cf: &str, key: &[u8], value: &[u8]) -> Result<()> {
+        if self.do_write(cf, key) {
+            let db = self.rocks.get_sync_db();
+            let handle = get_cf_handle(&db, cf)?;
+            return self
+                .rocks
+                .get_sync_db()
+                .put_cf(handle, key, value)
+                .map_err(r2e);
+        }
+        Ok(())
+    }
+
+    fn delete(&self, key: &[u8]) -> Result<()> {
+        if self.do_write(engine_traits::CF_DEFAULT, key) {
+            return self.rocks.get_sync_db().delete(key).map_err(r2e);
+        }
+        Ok(())
+    }
+
+    fn delete_cf(&self, cf: &str, key: &[u8]) -> Result<()> {
+        if self.do_write(cf, key) {
+            let db = self.rocks.get_sync_db();
+            let handle = get_cf_handle(&db, cf)?;
+            return self.rocks.get_sync_db().delete_cf(handle, key).map_err(r2e);
+        }
+        Ok(())
+    }
+
+    fn delete_range(&self, begin_key: &[u8], end_key: &[u8]) -> Result<()> {
+        // do nothing
+        Ok(())
+    }
+
+    fn delete_range_cf(&self, cf: &str, begin_key: &[u8], end_key: &[u8]) -> Result<()> {
+        // do nothing
+        Ok(())
+    }
+}
